[TCFMM-41] Move protocol fee tests into `Test.XToY`
Problem: Most of the tests in the `ProtocolFee.*` modules were
duplicated from `Test.XToY` and `Test.YToX`, and were modified to run
against the CTEZ variants of the contract. Now that the tests in
`Test.XToY` and `Test.YToX` are run against every variant, we no longer
need the tests in `ProtocolFee.*`.

Solution: Delete the `ProtocolFee.*` modules. There is no duplicate of
`test_protocol_fees_are_burned`, so we moved that specific test to
`Test.XToY`.
